# isme-nest-serrve.swagger.json
{
	"info": {
		"title": "isme-nest-serrve",
		"description": "",
		"version": "1.0.0"
	},
...
	"paths": {
		"/auth/login": {
			"post": {
				"summary": "登录",
				"deprecated": false,
				"description": "",
				"tags": ["认证"],
				"parameters": [{
					"name": "body",
					"in": "body",
					"schema": {
						"type": "object",
						"properties": {
							"username": {
								"type": "string"
							},
							"password": {
								"type": "string"
							},
							"captcha": {
								"type": "string"
							}
						},
						"required": ["password", "captcha", "username"]
					}
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "object",
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"required": ["accessToken"]
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [],
				"consumes": ["application/json"],
				"produces": ["application/json"]
			}
		},
...
		"/auth/current-role/switch/{roleCode}": {
			"post": {
				"summary": "切换角色",
				"deprecated": false,
				"description": "",
				"tags": ["认证"],
				"parameters": [{
					"name": "roleCode",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string",
					"x-example": "ROLE_QA"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "object",
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"required": ["accessToken"]
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/auth/logout": {
			"post": {
				"summary": "退出登录",
				"deprecated": false,
				"description": "",
				"tags": ["认证"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "boolean"
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
...
	},
	"swagger": "2.0",
	"basePath": null,
...
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
...
}